--- a/enterprise/server/tasksize/BUILD
+++ b/enterprise/server/tasksize/BUILD
@@ -21,10 +21,10 @@ go_library(
         "//server/util/authutil",
         "//server/util/log",
         "//server/util/perms",
+        "//server/util/proto",
         "//server/util/status",
         "@com_github_go_redis_redis_v8//:redis",
         "@com_github_prometheus_client_golang//prometheus",
-        "@org_golang_google_protobuf//proto",
     ],
 )